Disambiguate audio sample stride
Summary:
"Audio sample stride" has be used to designate the audio sample frame stride, meaning the stride of the group of audio samples captured synchronously, not the padded size of a single audio channel sample.
This diff changes the name of the constant and better documents what it means, without breaking compatibility with existing files, as the label saved in vrs files reamains unchanged.

Reviewed By: perdoch

Differential Revision: D64868230

fbshipit-source-id: 9ae1e85ccd1375f0812808cfcae84dfcd4bad2b9