More carefully handle TemporaryCachingStrategy
Summary:
TemporaryCachingStrategy is helper class that allows us to make a temporary caching strategy change. However, if the file handler has changed, it might lead to crashes.
This change makes sure we handle that situation properly, and only revert the caching strategy change is the handler is still the same.

Differential Revision: D68355971

fbshipit-source-id: d382b95d312db70cb0a3fe4f3995766fc6eed210