Allow FileHandler to point to a WriteFileHandler delegate
Summary:
Reading and writing from cloud storage are very different animals, and writing to the cloud has restrictions local files don't (writes are typically immutable, closing files typically involves concatenating uploaded chunks), which is why most FileHandlers can handle only read or write, but not both, except for DiskFile.
To make the process of writing a file to the cloud more transparent, and allow for URIs to not specify differents FileHandlers/URI schemes for read and write, we can allow reader FileHandlers point to their writer counterpart, that we can redirect to when a WriteFileHandler is requested.

Differential Revision: D68672737

fbshipit-source-id: af38b0d53c10399c6ed6112289481d3142e8c21a