Fix Unused Variable Error
Summary:
D68874718 causes some build issues (see https://www.internalfb.com/intern/test/562950120027900?ref_report_id=0 for an example) because the `inOutSpec` variable is not referenced in the function.

Adding the `[[maybe_unused]]` fixes this with the one test I ran, but I'll need CI to verify everything else. Let's use the macro `MAYBE_UNUSED` to be sure that this fix is supported by all compilers.

Differential Revision: D68985760

fbshipit-source-id: d3894c2e6c20055d6f512c16bd69dd47b7ac99e4