Make telemetry logger changes safer
Summary: Telemetry is mostly sent from background threads, so we need more care when changing the active telemetry logger, as it may be being called while it's changed. By delaying the deletion of loggers to static init teardown, happening after main() has returned, we reduce the chances that a background thread sends a message to a deleted logger. Note that having open files with active background threads when main() returns is still not supported.

Differential Revision: D68990357

fbshipit-source-id: 7b9891f94e4bf835fd0b84604718ded22265cfc7