Fix memory error on quit
Summary:
Users have reported a memory corruption on quit when working with some files. When I finaly was able to reproduce the issue in the debugger, I was able to see the error happened when destroying menus. When quitting, we end up updating the menus during teardown, which looks like it's the cause of the issues. By preventing menu updates during teardown, I can't repro the issue again.

note: this bug used to be fairly rare, but recent changes have made it much more frequent.

Differential Revision: D68998315

fbshipit-source-id: 7fc54436e49fac26fd8fabea2220fea192bf2455