Make sure to generate a CAT before quitting if we have telemetry to report
Summary:
Generating a CAT creates some telemetry we want to capture. But in some situations, this CAT creation happens after we disabled telemetry. This diff makes sure that we generate a CAT right before disabling telemetry, if there was something to log back.
Note that this isn't 100% bullet proof, but appears to catch the situations where the catch 22 was easy to observe.

Differential Revision: D71285750

fbshipit-source-id: f7dc65ac9fe93c09b4a96f132860a2524c3ca2d2