Make it possible to specify a type for custom content blocks
Summary:
Until now, custom content blocks were fully in the hands of the implementer, with, as only way to know what they might contain, recognize the combo recordable type id + record type + block index. This wasn't enough.

With this change, a custom content block can be given a format, specified as a string, with letters, numbers, and a number of special characters (but no space, slash, equal, plus).
To specify the format of a custom content block, use `CustomContentBlock("my_format")`

This allows you to easily specify what you store in your custom blocks when you create them, and you can now easily recognize what's in a custom content block when receiving them.

For instance, you could use "thrift:format:21" in configuration records, to specify a thrift format definition in the payload, and then "thrift:payload:21" in data records, for thrift payloads matching format 21. What you actually store in the content blocks, and knowing how to use the data, is still your responsibility, but now, it's much less ambiguous and easier to debug.

Reviewed By: hanghu

Differential Revision: D72098357

fbshipit-source-id: 824d46d623c47442ddc218a86507ff535b2f3ec8