Use paths as short as possible on windows (#176)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/176

Windows still have file path limitations (at least, some of the time), and we're seeing some tests fail because of it.
We've been using c:\tmp as a possible temp folder location on Windows, even though it's not an officially supported location, apparently with success.
So let's double-down on this, and try to use that path, but with some validation (trying to create a folder in it), otherwise, fallback to default APIs, while making sure we try to shorten the resulting path.

Differential Revision: D72607297