Make FileSpec::getEasyPath() more careful with delegators
Summary:
The uri field, when present in a FileSpec, is usually a trace of where the FileSpec came from: it's probably the uri that was parsed to populate the fields of the FileSpec.  But it might no longer be accurate if the FileSpec was manipulated directly unceremoniously afterwards.

If the presence of a delegator not visible, this can cause confusion as getEasyPath() is typically used for logging, on error in particular.

With this diff, when using the uri field for getEasyPath(), should there be a delegator in the FileSpec not directly visible in the uri, we make sure to reveal the delegator.

Reviewed By: maxouellet

Differential Revision: D72614625

fbshipit-source-id: 632a7bb93c32c9662f78583e891a85c9df344878