Add --json-pretty-description option
Summary:
The "json-description" command in vrstool and the OSS' vrs cli tool produce a long json string that's impossible for humans to read, because it's designed to be read by code. However, some folks would like to read this json, maybe so they can write a parser for it, or maybe they like the presentation better than the default vrs file information presentation these tools provide.
This diff adds a "json-pretty-description" command that does this.

Differential Revision: D73020708

fbshipit-source-id: a8dfbcbcaa4b06f37fce3c622d04ffa9b7a103a8