Fix rapidjson string handling issue
Summary:
RapidJson agressively uses the optimization that strings don't have to be copied, which can lead to bugs. They also have issues when copying objects, which, for reasons unclear, started to break on Windows only recently.
This diff provides a safe and unified way to specify which method should be used, rather than relying on passing an allocator to mean to copy the string data. It also removes some definition duplications.

Differential Revision: D73126297

fbshipit-source-id: ba05409d85b5ab60d048c2db24ac17f5b73cbc30