Add support for boost::process v1.88.0 (#180)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/180

Boost/process has API breaking changes starting with version 1.88.0 created less than 2 weeks ago in open source and that brew is starting to offer. Starting with boost v1.88.0, process/v2 is the default.
We can't switch to process/v2, because we still need to support versions of boost that don't include any version of it. The good news is that boost 1.88.0 still includes process/v1, even if you have to go find it. The new <vrs/os/Process.h> header helps map the includes and definitions, so we can more easily continue to use boost/v1 until we can do a full switch to v2.

Reviewed By: YLouWashU

Differential Revision: D73412134

fbshipit-source-id: bec7905d10de163de5f19ad3a9645c8da63f9644