Hide ProgressLogger from Fast-DDS
Summary:
Fast-DDS is an open source library with horrifying defines in their headers colliding with some of our API when the two libraries are used together. Shame on them.

From a practical standpoint, it is cleaner and safer to avoid the collision rather than fighting defines with undefines, which can cause further unintended consequences, in particular because we don't use the affected symbol in vrs. So it's easy and safer to avoid that name, and we can avoid exposing this header, too, in case something else happens.

Reviewed By: lucasgadani

Differential Revision: D73488457

fbshipit-source-id: 355510f5a06ca876ff5e44febd777a836d9e9148