Make sure to instantiate all required templates
Summary:
When template classes are not fully defined in headers to reduce burden (compile time) on all consumers of the code, the methods need to be instantiated explicitly for each type used. This method is commonly used for DataLayout which only allows specific data types.
For some reason, this hasn't been been a problem until now, but some linkage issues have poped-up that this change should address.

Reviewed By: vener91

Differential Revision: D73513268

fbshipit-source-id: 45af9f34e8b291c285117c01ce9cf209a498a36d