Fix sanity check edge case
Summary:
When a VRS file has no record, its index record might not have passed the record header sanity check, making it appear as if the file was corrupt, when it was actually healthy.

The updated check:
- properly accounts for the record header size
- is more or less demanding, depending on the size of the payload, being less demanding for smaller payloads.

Note that for a little while, we were not systematically validating that compression was helping, possibly leading to records slightly larger than their uncompressed original, so we can't require that compressed payloads are actually smaller than the uncompressed data, even for large records.

Reviewed By: finik

Differential Revision: D73702998

fbshipit-source-id: bae2f22ef2e67a57c4c84821b78785c50d555055