Add methods to copy values from a datalayout to another
Summary: It can be useful to copy values found in an existing `DataLayout` into another `DataLayout`. This can be an expensive operation, therefore, this method requires the use of mapped layout probably provided by the `getExpectedLayout()` method, so the copy can leverage a cached mapping and be as fast as possible.

Differential Revision: D73756774

fbshipit-source-id: 281ebbc3f7f046938a59a50cf1cbb0fe510941fa