Provide options to replace datalayout entirely in doDataLayoutEdits() callbacks
Summary:
In some extreme datalayout edit situations, users may want to replace the datalayout with an entirely new definition.
Users can now do that by calling pushDataLayout() directly with the new dalalayout. However, the filter needs to detect that pushDataLayout() was already called and not push the "old" datalayout, which would make the record have two datalayout instead of one.

Reviewed By: finik

Differential Revision: D73828851

fbshipit-source-id: 3d529bb435cabbfd847786078775a583b4739ba5