Add method to explicitly init a DataLayout and DataPieces to their default value
Summary: DataPiece default values are primarily intended to be returned when trying to get the value of a data piece that isn't mapped (that is, found in a datalayout read from a file). Until now, this was their only use. With this dfif, we make it possible to intialze (set or stage) all the data pieces of a datalayout with their default value.

Differential Revision: D73898069

fbshipit-source-id: 3964b09336d085bf106becdd7aa77a3383b18aae