Add shell scripts to build FFmpeg libs.
Summary:
This diff adds a shell script for users to download and install FFmpeg library, which is a prerequisite for XPRS decoding feature in VRS repo. This shell script would install FFmpeg under `$HOME/vrs_third_party_libs/ffmpeg/` folder. This path is later used by `FindFFmpeg.cmake` file to link to this library.

The ffmpeg lib is installed with the following flags to avoid license complications:
```
./configure --prefix="$INSTALL_PATH" --disable-swresample --disable-swscale --disable-everything --enable-decoder=hevc --enable-zlib --enable-static --disable-shared
```

Reviewed By: georges-berenger

Differential Revision: D73452900

fbshipit-source-id: 82c65bb48d19cd51b6f755e439e05f1238fc8c5a