Add CMAKE find_package for FFmpeg libs.
Summary: This diff adds a `FindFFmpeg` module to vrs's cmake folder. This lib is only used by the decoding feature, therefore not declared as `REQUIRED` in cmake. Users should call the shipped `build_ffmpeg_linuxunix.sh` script to install ffmpeg to a specified location in order for this to work.

Reviewed By: georges-berenger

Differential Revision: D73452897

fbshipit-source-id: e0f626dbc4b46da7de8f7717490ef11681aa865c