Fixing FindOcean.cmake cannot run twice issue. (#190)
Summary:
Pull Request resolved: https://github.com/facebookresearch/vrs/pull/190

In D68536279, we ran into an issue where after introducing Ocean lib to vrs repo, running ` cmake -S . -B ./build/ -G Ninja ` multiple times would emit an error on `Ocean::Ocean` not found, we had to remove the `build/` directory in order to run it again. And we added a `clean` step in Pixi as a workaround of this issue.

I found the root cause is because we cache some Ocean related variables in `FindOcean.cmake`, causing a second build to fail. This diff removed the cached variables that causes the issue.

Reviewed By: georges-berenger

Differential Revision: D74092912

fbshipit-source-id: 580e67ca8a228c0f0af66595a92f4ac284aac125