Make it simpler to work with stream filters
Summary:
Stream filters can be used in different ways. The original way was to control which streams are copied, which are not, but the same logic/syntax can be used in the context of filters to control which streams are modifed which are copied as is.

This helper method helps avoid cloning code as this pattern starts to be repeated.

Differential Revision: D74847033

fbshipit-source-id: e9e43c516e7ebde48752b033654db76a203c29ee