introduce conversion MultiRecordFileReader -> RecordFileReader
Summary:
In some cases, it could be convenient to take a MultiFileRecordReader and cast it to a RecordFileReader.
Mainly to interface becode code that needs to handle MultiFile and code that does not use MultiFileRecordReader (yet)

This conversion can be done when the MultiFileRecordReader actually contains one file.

This diff propose and helper methods that does this conversion or crashes. The MultiRecordFileReader is swallowed in the function to avoid having around multiple reference.

Reviewed By: georges-berenger

Differential Revision: D75457291

fbshipit-source-id: 2bc36841a544ef9d48444b4f652f3a26c3e2268a