Split AsyncDiskFile.hpp into .cpp/.h files
Summary: AsyncDiskFileChunk is pretty heavy, with significant windows-only and linux-only parts. It also uses EnumStringConverter in a way that doesn't sit well with ms compilers and prevents the next code refactor. By splitting the file in a more conventional .cpp/.h pair, we can make it easier to compile. No behavior change is expected.

Differential Revision: D77237156

fbshipit-source-id: 326ea6524ae0344fcdc8e6029937e212d540d92b