Add option to control if the index record should be pre-allocated
Summary:
By default, during filter-copy operations, when the output file is created locally, vrs pre-allocates space for the index record early in the file. While this usually makes the file more streaming-friendly if it's later uploaded to the cloud, in some situations, the index can be significantly smaller or bigger than anticipated, causing some file size "surprises". This is usually not a problem (the files are fine and the "surprises" are at most 10 to 20 MB large when dealing with large indexes), but it can create confusion, in particular when trying experimenting with data size.
This new option gives control, so we can avoid this condition altogether.

Note that these conditions never happen when a file is created in the cloud, because of how the index can be prepended to the file, at the end of the file's head.

Differential Revision: D77634889