Isolate default decimator implementation
Summary:
The Decimator code is really tied to vrstool and its vrs CLI open source counterpart.
This diff isolates it in a new DefaultDecimator class that derives from an abstraction that can be better reused. That implementation belongs with cli utils, including `DecimationParams`, the original controls set from CLI tools.

Reviewed By: jeffreyliu3230

Differential Revision: D77642517

fbshipit-source-id: f2fe4741c78712522af23319ee3cae39aff07bd6