Add streamID and data record index to direct image references
Summary:
This adds stream context to direct image references, so that we now know in which stream an image belongs, and its data record index.
Note: in the text index file generated, we're replacing a simple counter that's not actually very useful (it's not an index usable in the vrs file), with the streamID + data record index in that stream.

Reviewed By: maxouellet

Differential Revision: D77770762

fbshipit-source-id: e9e1ce3e1baaee5c27d2a543e9faff74bd283309