Fix inconsistencies with return value of DataPieceString::get(var) and default values
Summary:
This diff addresses inconsistencies when handling `bool DataPieceXXX::get(var)` calls, and to match documentation:
- if a field is available, set explicitly or not, `get()` should return true, no matter the value.
- when a field itself is unavailable, default values are returned, but only then.

A field is unavailable when a layout is mapped, but the field isn't found in the target layout. This can happen when a field is added to a `DataLayout`, and an old file is read that doesn't contain the new field in its `DataLayout`.

In particular, this diff addresses variable size fields, for which an empty value, meaning, an empty string, an empty vector, or an empty map, would be be confused with an unavailable field, so that the passed `var` argument would be set to the default value, and false returned.

Note: Fixing `DataPieceVector<T>` and `DataPieceStringMap<T>` exposes a few tests that need to be addressed, so this diff does not attempt to fix these types of fields: see next diff.

Reviewed By: maxouellet

Differential Revision: D78055666

fbshipit-source-id: b6cc39f2736098c6dae9b1a0150950d0504d478b