Fix inconsistencies with return value of DataPieceVector<T>::get(var), DataPieceStringMap<T>::get(var) and default values"
Summary:
This diff addresses inconsistencies when handling `bool DataPieceVector::get(var)` and `bool DataPieceStringMap::get(var)` calls, and to match documentation:
* if a field is available, set explicitly or not, `get()` should return true, no matter the value, including an empty vector.
* when a field itself is unavailable, default values are returned, but only then.

A field is unavailable when a layout is mapped, but the field isn't found in the target layout. This can happen when a field is added to a DataLayout, and an old file is read that doesn't contain the new field in its DataLayout.

In particular, this diff addresses variable size fields, for which an empty value, meaning, an empty vector would be be confused with an unavailable field, so that the passed var argument would be set to the default value, and false returned.

Differential Revision: D78082013

fbshipit-source-id: a1737b8532755914452d80e72225bdc2d2641dad