Edit ImageSupport.md using inpage editor
Summary:
Image support updates to the vrs open source documentation, to reflect open sourcing xprs and the addition of  `image/jxl` and `image/custom_codec`.

NOTE: If you want to update this diff, go via the preview link inside the static docs section below.
Ensure you are editing the same page that was used to create this diff.

Differential Revision: D78191179

fbshipit-source-id: e68c2ee5f2efb70ae8cc63f33aa5f915aad0c0b9