Make PrefetchCacheFileHandler usable through URI parameter
Summary:
This diff makes `PrefetchCacheFileHandler` usable through a URI parameter.
The changes makes `CachedFileHandler`, the default class that handles caching for network FileHandlers, able to delegate to `PrefetchCacheFileHandler` if the kPrefetchCache extra is set in the file spec.

These changes allow for prefetching to be enabled or disabled on a per-file basis using a URI parameter.
To use `PrefetchCacheFileHandler`, add the URI parameter `?prefetch_cache=1` to Gaia or Manifold URIs.

Reviewed By: tschmidt23

Differential Revision: D79421211

fbshipit-source-id: f561c978551a46906aad4c1ada010e67690d35aa