Add resize options to CompressAsyncImageFilter
Summary: This diff adds resize options to the CompressAsyncImageFilter class in the vrs library. The resize options allow for downscaling or upscaling of images during compression. The changes include adding a new struct for resize options, modifying the constructor of the CompressionJob class to include the resize options, and modifying the CompressAsyncImageFilter class to use the resize options.

Differential Revision: D79710550

fbshipit-source-id: b2abcd1989ef0a67f090a7623654a1f2dc78c776