Add method to specify related tags in MultiRecordFileReader
Summary:
This diff adds a method to specify related tags in the MultiRecordFileReader class.
The setRelatedFileTags() method is added to allow users to set their own related file tags.
The new relatedFileTags_ member variable is initialized with the default values of kCaptureTimeEpoch and kSessionId, to preserve current behaviors.
The MultiRecordFileReaderTest.cpp file is modified to test the new methods.

Differential Revision: D80096296

fbshipit-source-id: b0beaea4f13731714f90806e3f41962c8f9800a5