More conversion fixes & validation
Summary:
Some combo of input pixel formats, normalize and resize options were still not handled properly. In particular, some streams were reported as processed when there were no changes to be made, and controverselly, some streams were not processed that should have been, like the example in the test plan.

So this diff revisits the logic from scratch, and adds a number of deliberate point checks that were manually reviewed for accuracy (NO AI INVOLVED IN THE LOGIC).
The problem, is that AI easily just makes tests that pass with the logic of the tested code, as opposed to validating the tested code. AI is awesome at writing tests that validate nothing, and needs more supervision than a 3 year old in a candy store.

The only way to avoid the trap is to hand-write validation for the combos you really care about and not allow AI to write that code for you.

Differential Revision: D80485713

fbshipit-source-id: 31b073c7b1ec9102423eaf3b0be54547b6ceb8ff