Use more enum template helpers
Summary:
Enums can be a pain to work with in C++, and these helper templates help with classic operations, as long as the enum respects some conventions.
With this diff, we add more templates that expands the existing templates so we can iterate over the valid enum values, and validate that an enum value is valid.

No functional changes are expected, but the code looks better!

Differential Revision: D80567566

fbshipit-source-id: b6d28cb0c7c218c17912027bf18596d465a33adf