Exposing xprs + ffmpeg libraries as PUBLIC in CMake. (#208)
Summary:

This diff exposes the xprs and ffmpeg libraries as PUBLIC in CMake. This is needed by the Nebula ClientSDK project, where it needs to pybind some XPRS functions within projectaria_tools_gen2 in D79454506, and currently all XPRS + FFmpeg libs are not visible there, hence the build CI error in that diff.

Reviewed By: georges-berenger, danielyan86129

Differential Revision: D80776108