Improve image data error reporting
Summary:
A common bug in recording apps is to same a number of bytes that doesn't match the image spec. This ends up being reported as a bug in other places, including vrsplayer.
This diff adds a bit more proactive detection of these errors with more explicit reporting and finger pointing.

Reviewed By: finik

Differential Revision: D82601779

fbshipit-source-id: d56809e908a8cd92c3b49aa7d9474a991f14b5fe