Remove LittleEndian<T>
Summary:
All the systems vrs supports are little endian, and there are many other little endian dependencies all over the place that would be really difficult to remove. Therefore, we should not pretend that vrs could easily be made to work on a little endian system and the LittleEndian<T> class should be removed. This class also makes debugging more painful, as the values are all hidden in containers.

So everywhere `LittleEndian<T> variableName` is used, we replace it with a simple `T variableName`, then fix all the places variableName is used with direct manipulation of the member variable without going through any getter or setter. By adding operators in previous diff, we reduced the number of places where we still use getters to a very small number.

LittleEndian<T> used to default init its value, so we need to carefully init these variables to avoid introducing bugs.

Differential Revision: D83631395

fbshipit-source-id: 8dcd133a0832cc201068419ceb133380edbc6d84