Add miscParams field to CopyOptions for custom parameters
Summary:
CopyOptions contains a number of special fields to specify options used during copy operations, and this struct is passed in many places. However, we don't want to keep augmenting it, in particular, we don't want to add filter specific options to it. But since it's a very convenient place to place custom parameters, let's just add a map<string, string> to it, as we've been using throughout vrs, FileSpec in particular, to hold custom parameters.

Note: this diff is isolated because this file is open sourced. The new parameter isn't used here.

Differential Revision: D84963867

fbshipit-source-id: a14c1e8b5a8c67f5f4c8211993a836a34c1f8601