Update headers to support raw_headers_as_headers_enabled
Summary:
This updates `#include` paths for headers that are inherited from the `stream_identifiers` target to reference them only from the public include directory to remove the implicit assumption that its source header file location matches the downstream `vrs` Buck target's as that breaks Clang header maps that we want to enable in the next diff D85594489.

This revealed that a number of them were using the incorrect root folder for such include paths even when not using file-relative paths as well as the include directories not always being correctly defined in the BUCK files which were both also fixed.

Also updated neighboring `#include`s in the changed files for better consistency.

Reviewed By: georges-berenger

Differential Revision: D85594490

fbshipit-source-id: b7f56c3a6c4d8a2ed0442aef454b1ac646c0aa48